fix: Don't block shutdown with monitor thread (#285)
This exists purely to signal minidump crash events, which are currently
not enabled in the latest release. Get rid of the infrastructure for
handling it via pipes, and we'll use `WaitForMultipleObjects` on the
process/crash event.

Additionally switch the mod host to logging to `stdout`, so we capture
logs from any other DLL mods in use.

Fixes #270.