fix: Launch path normalization (#365)
If you just point directly to a file in the command line with the
profile like so `me3.exe launch -p mods.me3 --auto-detect`, the parent
of `mod.me3` will just be an empty path, which `normalize` refuses to
handle.

My quick solution for this is to just normalize first, then take the
parent of the normalized path.